Fix Betty codestyle errors

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -11,14 +11,14 @@ void print_number(int n)
 
 	n = (long) n;
 
-	if ( n < 0)
+	if (n < 0)
 	{
 		_putchar('-');
 		n = -n;
 
-	for (i = (int) 10e8; i >= 1; i /= 10)
+	for (i = 10e8; i >= 1; i /= 10)
 	{
-		if ((n / i) != 0 )
+		if ((n / i) != 0)
 			_putchar((n / i) + '0');
 		n = n % i;
 	}
